requirements.txt
fastapi
uvicorn
pydantic
pydantic_settings
qdrant-client
sentence-transformers
aio_pika
pyjwt
torch==2.3.1+cpu

-f https://download.pytorch.org/whl/torch_stable.html
